pyproject.toml
[tool.poetry]
name = "cauch-e"
version = "0.1.0"
description = ""
authors = ["Cyclic3 <cyclic3.git@gmail.com>"]
readme = "README.md"
packages = [{include = "cauch_e"}]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
pyyaml = "^6.0"
inquirer = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
